+++ src/UC_Quantum_Lab/_src.py
@@ -16,7 +16,7 @@ def _trigger():
 prepends inputted path with the config directory if it exists
 
 """
-def get_path(path:str):
+def _get_path(path:str):
     global _config_dir
     if _config_dir in os.listdir(): 
         if os.path.isdir(_config_dir): return os.path.join(_config_dir, path)