+++ src/UC_Quantum_Lab/commands.py
@@ -5,13 +5,13 @@ import matplotlib.pyplot as plt
 from atexit import register
 from math import log
 from . import _states, _circs, _hists, _master_show, _show_plt
-from ._src import get_path
+from ._src import _get_path
 
 _circ_count = 0
 _state_count = 0
 _hist_count = 0
 
-def message(msg):
+def _message(msg):
     print(f"UC_Quantum_Lab: {msg}")
 
 def _show_at_exit():
@@ -23,16 +23,16 @@ def _show_at_exit():
 register(_show_at_exit)
 
 # diplays the image in the viewer or saves the image to the inputted path
-def display(circuit:QuantumCircuit, path:str=""):
+def display(circuit:QuantumCircuit, path:str="")->None:
     global _circ_count, _circs, _master_show, _show_plt
     circuit.draw(output='mpl')
     plt.tight_layout()
     if len(path): 
-        message(f"outputing circuit diagram to \"{path}\"")
+        _message(f"outputing circuit diagram to \"{path}\"")
         plt.savefig(path)
     elif _master_show:
         #print("displaying circuit")
-        p = get_path(f"_circ_{_circ_count}.png")
+        p = _get_path(f"_circ_{_circ_count}.png")
         plt.savefig(p)
         _circs.append(p)
         _circ_count+=1
@@ -40,18 +40,18 @@ def display(circuit:QuantumCircuit, path:str=""):
         _show_plt = True
 
 # generates binary strings
-def getbin(n, s=['']):
+def _getbin(n, s=['']):
     global _config
-    if n > 0: return [*getbin(n - 1, [i + '0' for i in s]), *getbin(n - 1, [j + '1' for j in s])]
+    if n > 0: return [*_getbin(n - 1, [i + '0' for i in s]), *_getbin(n - 1, [j + '1' for j in s])]
     return s
 
 # displays the statevector of the circuit and can return it
-def state(circuit:QuantumCircuit, show=True):
+def state(circuit:QuantumCircuit, show:bool=True)->list[complex]:
     global _state_count, _states, _master_show, _show_plt
     _state = Statevector.from_instruction(circuit).data
     _num_bits = int(log(len(_state))/log(2))
     
-    _options = getbin(_num_bits)
+    _options = _getbin(_num_bits)
     _data = {}
     for i in range(len(_state)):
         val = _state[i]
@@ -64,7 +64,7 @@ def state(circuit:QuantumCircuit, show=True):
     if show and _master_show:
         #print("showing state vector")
         if len(_states):
-            if len(_options[i]) > len(list(_states.keys())[0]):
+            if len(_options[-1]) > len(list(_states.keys())[0]):
                 raise KeyError("States must be obtained from the same circuit")
             for item in _data:
                 _states[item].append(_data[item])
@@ -75,22 +75,29 @@ def state(circuit:QuantumCircuit, show=True):
     return _state
 
 # displays the histogram of the circuit after execution in the viewer
-def counts(circuit:QuantumCircuit, backend=Aer.get_backend('qasm_simulator'), path:str=""):
+def counts(circuit:QuantumCircuit, backend=Aer.get_backend('qasm_simulator'), path:str="", show:bool=True) -> dict[str, int]:
     global _hist_count, _hists, _master_show, _show_plt
     counts = execute(circuit, backend=backend, shots=1024).result().get_counts()
+    # accounting for weirdness of counts result from qiskit
+    for item in list(counts):
+        new_item = item.split(" ")[0]
+        counts[new_item] = counts[item]
+        del counts[item]
+
     if len(path): 
-        message(f"outputing histogram to \"{path}\"")
+        _message(f"outputing histogram to \"{path}\"")
         plt.savefig(path)
-    elif _master_show:
+    elif _master_show and show:
         #print("displaying histogram")
         plot_histogram(counts)
-        p = get_path(f"_hist_{_hist_count}.png")
+        p = _get_path(f"_hist_{_hist_count}.png")
         plt.savefig(p)
         _hists.append(p)
         _hist_count+=1
-    else:
+    elif show:
         plot_histogram(counts)
         _show_plt = True
+
     return counts